# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProSAFE Plus NSDP API (LLM configuration interface)",
    "version": "0.1.0",
    "description": "Local HTTP API for managing NETGEAR ProSAFE Plus switches over NSDP, designed as an LLM tool-calling / agent configuration interface. `prosafe serve` exposes this OpenAPI document at /openapi.json so a language model (or any generated client) can drive the switch: discover devices, log in, then read and write every configurable page. Each operation has a stable operationId suitable for use as a tool name. Workflow for an agent: 1) POST /api/discover, 2) POST /api/switch/{mac}/login with the password, 3) call the GET endpoints to read state and the POST endpoints to change it. Login state is held per server process. Destructive operations (reboot, factory-reset, firmware) are real and immediate. Start the server with --token to require an X-Auth-Token header.",
    "license": {
      "name": "MIT"
    },
    "x-intended-for": "LLM-based configuration / agent tool-calling"
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Token",
        "description": "Required only when the server was started with --token."
      }
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "token": []
    }
  ],
  "paths": {
...
    "/api/devices/add": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Add a switch by IP (across subnets)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ip"
                ],
                "properties": {
                  "ip": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceRow"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "502": {
            "$ref": "#/components/responses/SwitchError"
          }
        },
        "operationId": "do_devices_add"
      }
    },
    "/api/switch/{mac}/login": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Authenticate to a switch",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "operationId": "do_login"
      }
    },
    "/api/switch/{mac}/info": {
      "get": {
        "tags": [
          "info"
        ],
        "summary": "Switch information",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "responses": {
          "200": {
            "description": "info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        },
        "operationId": "get_info"
      }
    },
    "/api/switch/{mac}/name": {
      "post": {
        "tags": [
          "info"
        ],
        "summary": "Set switch name",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_name"
      }
    },
    "/api/switch/{mac}/ip": {
      "post": {
        "tags": [
          "info"
        ],
        "summary": "Set IP settings (DHCP or static)",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IPSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_ip"
      }
    },
    "/api/switch/{mac}/ports": {
      "get": {
        "tags": [
          "ports"
        ],
        "summary": "Port link status",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "responses": {
          "200": {
            "description": "ports",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortsResponse"
                }
              }
            }
          }
        },
        "operationId": "get_ports"
      }
    },
    "/api/switch/{mac}/ports/admin": {
      "post": {
        "tags": [
          "ports"
        ],
        "summary": "Set a port's speed and flow control",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "port": {
                    "type": "integer"
                  },
                  "speed": {
                    "type": "integer"
                  },
                  "flow_control": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_ports_admin"
      }
    },
    "/api/switch/{mac}/stats": {
      "get": {
        "tags": [
          "ports"
        ],
        "summary": "Port statistics",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "responses": {
          "200": {
            "description": "stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortStats"
                  }
                }
              }
            }
          }
        },
        "operationId": "get_stats"
      }
    },
    "/api/switch/{mac}/stats/clear": {
      "post": {
        "tags": [
          "ports"
        ],
        "summary": "Clear all port counters",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "do_stats_clear"
      }
    },
    "/api/switch/{mac}/cabletest": {
      "post": {
        "tags": [
          "ports"
        ],
        "summary": "Run the cable tester on ports",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ports"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CableResult"
                      }
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "do_cabletest"
      }
    },
    "/api/switch/{mac}/mirror": {
      "get": {
        "tags": [
          "ports"
        ],
        "summary": "Get mirroring config",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "responses": {
          "200": {
            "description": "mirror",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mirror"
                }
              }
            }
          }
        },
        "operationId": "get_mirror"
      },
      "post": {
        "tags": [
          "ports"
        ],
        "summary": "Set mirroring (dest 0 disables)",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mirror"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_mirror"
      }
    },
    "/api/switch/{mac}/vlan": {
      "get": {
        "tags": [
          "vlan"
        ],
        "summary": "VLAN configuration",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "responses": {
          "200": {
            "description": "vlan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VLANResponse"
                }
              }
            }
          }
        },
        "operationId": "get_vlan"
      }
    },
    "/api/switch/{mac}/vlan/mode": {
      "post": {
        "tags": [
          "vlan"
        ],
        "summary": "Set VLAN engine mode (0-4; resets membership)",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mode": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 4
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_vlan_mode"
      }
    },
    "/api/switch/{mac}/vlan/port": {
      "post": {
        "tags": [
          "vlan"
        ],
        "summary": "Set port-based VLAN membership",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortVLAN"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_vlan_port"
      }
    },
    "/api/switch/{mac}/vlan/8021q": {
      "post": {
        "tags": [
          "vlan"
        ],
        "summary": "Create or update an 802.1Q VLAN",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VLAN8021Q"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_vlan_8021q"
      }
    },
    "/api/switch/{mac}/vlan/8021q/delete": {
      "post": {
        "tags": [
          "vlan"
        ],
        "summary": "Delete an 802.1Q VLAN",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vid": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_vlan_8021q_delete"
      }
    },
    "/api/switch/{mac}/vlan/pvid": {
      "post": {
        "tags": [
          "vlan"
        ],
        "summary": "Set the PVID on ports",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ports": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "vid": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_vlan_pvid"
      }
    },
    "/api/switch/{mac}/qos": {
      "get": {
        "tags": [
          "qos"
        ],
        "summary": "QoS, rate limits and broadcast filtering",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "responses": {
          "200": {
            "description": "qos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QoSResponse"
                }
              }
            }
          }
        },
        "operationId": "get_qos"
      }
    },
    "/api/switch/{mac}/qos/mode": {
      "post": {
        "tags": [
          "qos"
        ],
        "summary": "Set QoS mode (1 port-based, 2 802.1p)",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mode": {
                    "type": "integer",
                    "enum": [
                      1,
                      2
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_qos_mode"
      }
    },
    "/api/switch/{mac}/qos/priority": {
      "post": {
        "tags": [
          "qos"
        ],
        "summary": "Set port priority (1 high - 4 low)",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ports": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "priority": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 4
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "set_qos_priority"
      }
    },
...
    "/api/switch/{mac}/reboot": {
      "post": {
        "tags": [
          "maintenance"
        ],
        "summary": "Reboot the switch",
        "parameters": [
          {
            "$ref": "#/components/parameters/mac"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OK"
          }
        },
        "operationId": "do_reboot"
      }
    },
...
  }
}